% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base64_to_img}
\alias{base64_to_img}
\title{Image data uri to file}
\usage{
base64_to_img(img64, slug)
}
\arguments{
\item{img64}{charachter, base64 image string as returned by \link{generate_txt2img}}

\item{slug}{character, name of file to export image to. WITHOUT extension}
}
\value{
nothing, called for side effects
}
\description{
Convert a data uri to an image in the correct format and save it to a file.
}
