% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{generate_txt2img}
\alias{generate_txt2img}
\title{Text to Image
Create an image from a text prompt}
\usage{
generate_txt2img(
  text_prompts = "",
  engine_id = "stable-diffusion-xl-1024-v0-9",
  Accept = "application/json",
  token = NULL,
  ...
)
}
\arguments{
\item{text_prompts}{string.}

\item{engine_id}{string.}

\item{Accept}{string. One of "application/json" "image/png".}

\item{token}{either the stability.ai API key or NULL to read it from .Renviron}

\item{...}{further parameters to pass to the API}
}
\value{
png image as base64. can be saved with \link{base64_to_img}
}
\description{
Text to Image
Create an image from a text prompt
}
\details{
for a detailed list of supported parameters see \url{https://platform.stability.ai/docs/api-reference#tag/v1generation/operation/textToImage}
}
\examples{
\dontrun{
generate_txt2img(text_prompts = "A lighthouse on a cliff")
}
}
