% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_account_balance}
\alias{get_account_balance}
\title{Account balance
get balance from your stability.ai account}
\usage{
get_account_balance(token = NULL)
}
\arguments{
\item{token}{either the stability.ai API key or NULL to read it from .Renviron}
}
\value{
account balance
}
\description{
Account balance
get balance from your stability.ai account
}
\examples{
\dontrun{
get_account_balance()
}
}
