% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_account_details}
\alias{get_account_details}
\title{Account details
get details about your stability.ai account}
\usage{
get_account_details(token = NULL)
}
\arguments{
\item{token}{either the stability.ai API key or NULL to read it from .Renviron}
}
\value{
data frame with account details
}
\description{
Account details
get details about your stability.ai account
}
\examples{
\dontrun{
get_account_details()
}
}
