% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_engines_list}
\alias{get_engines_list}
\title{Engines list
get a list of engines to use}
\usage{
get_engines_list(token = NULL)
}
\arguments{
\item{token}{either the stability.ai API key or NULL to read it from .Renviron}
}
\value{
data.frame containing engines
}
\description{
Engines list
get a list of engines to use
}
\examples{
\dontrun{
get_eninges_list()
}
}
